task

+++ src/main/java/com/app/service/ImageService.java
@@ -1,25 +1,27 @@
 package com.app.service;
 
+import com.app.dto.GetImagesResponse;
+import com.app.dto.ImageResponse;
 import com.app.dto.UploadImageResponse;
 import com.app.exception.ImageIsTooBigException;
+import com.app.exception.ImageNotAccessibleException;
 import com.app.exception.ImageNotFoundException;
 import com.app.exception.NotSupportedTypeOfImageException;
-import com.app.mapper.ImagesMapper;
 import com.app.model.ImageMeta;
+import com.app.model.user.User;
 import com.app.repository.ImageMetaRepository;
+import com.app.service.auth.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisHash;
 import org.springframework.data.util.Pair;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
-@RedisHash
 @RequiredArgsConstructor
 public class ImageService {
 
@@ -33,11 +35,19 @@ public class ImageService {
 
     private final MinioService imageStorage;
 
+    private final UserService userService;
+
     public void deleteImage(String imageId) throws Exception {
         Optional<ImageMeta> imageMeta = repository.findImageMetaByImageId(imageId);
         if (imageMeta.isEmpty()) {
             throw new ImageNotFoundException();
         }
+
+        User user = userService.getCurrentUser();
+        if (!Objects.equals(imageMeta.get().getUserId(), user.getId())) {
+            throw new ImageNotAccessibleException();
+        }
+
         imageStorage.deleteImage(imageId);
         repository.deleteById(imageMeta.get().getId());
     }
@@ -48,10 +58,15 @@ public class ImageService {
             throw new ImageNotFoundException();
         }
 
+        User user = userService.getCurrentUser();
+        if (!Objects.equals(imageMeta.get().getUserId(), user.getId())) {
+            throw new ImageNotAccessibleException();
+        }
+
         return Pair.of(imageStorage.downloadImage(imageId), imageMeta.get().getMediaType());
     }
 
-    public UploadImageResponse uploadImage(MultipartFile file) throws Exception {
+    public String uploadImage(MultipartFile file) throws Exception {
         if (file.getSize() > MAX_ALLOWED_IMAGE_SIZE) {
             throw new ImageIsTooBigException(file.getSize(), MAX_ALLOWED_IMAGE_SIZE);
         }
@@ -59,14 +74,22 @@ public class ImageService {
             throw new NotSupportedTypeOfImageException();
         }
 
+        User user = userService.getCurrentUser();
+
         String imageId = imageStorage.uploadImage(file);
         repository.save(new ImageMeta()
                 .setImageId(imageId)
                 .setName(file.getOriginalFilename())
                 .setSize(file.getSize())
                 .setMediaType(file.getContentType())
+                .setUserId(user.getId())
         );
 
-        return new UploadImageResponse(imageId);
+        return imageId;
+    }
+
+    public List<ImageMeta> getImages() {
+        User user = userService.getCurrentUser();
+        return repository.findAllByUserId(user.getId());
     }
 }
